# Perf - don't use do block for readtable