This ends up being surprisingly expensive!  

My timings include pending PRs to EzXML & ZipFile - for reference pandas parses this file in ~80 ms, and R's read_xl, which has a better parser, does it in ~45ms, so post this + those changes pretty happy with where overall performance is.

```
# master
julia> @btime XLSX.readtable(f, "Sheet1")
  112.970 ms (280165 allocations: 11.58 MiB)

# pr
julia> @btime XLSX.readtable(f, "Sheet1")
  47.748 ms (280068 allocations: 11.58 MiB)
```