@felipenoris 2020-05-31
@chris-b1 thanks for finding this! I think bd6f607 will do the job. If not, please open an issue. Thanks!